Rewrite and fixup of user/auth related modules
