+++ package.json
@@ -57,7 +57,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.45.2",
-    "react-native": "0.73.4",
+    "react-native": "0.73.6",
     "react-native-gesture-handler": "~2.14.0",
     "react-native-image-colors": "^2.4.0",
     "react-native-ios-context-menu": "^2.4.3",
