+++ src/store/index.ts
@@ -1,15 +1,9 @@
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 import type { PreloadedState } from '@reduxjs/toolkit';
 
-import { spotifyApi } from 'services/spotify.api';
-import { supabaseApi } from 'services/supabase.api';
-import { userListener } from 'store/listenerMiddleware';
-
 import userReducer from 'user/user.slice';
 
 const rootReducer = combineReducers({
-  [supabaseApi.reducerPath]: supabaseApi.reducer,
-  [spotifyApi.reducerPath]: spotifyApi.reducer,
   user: userReducer,
 });
 
@@ -18,9 +12,7 @@ export const setupStore = (preloadedState?: PreloadedState<RootState>) => {
     reducer: rootReducer,
     preloadedState,
     middleware: (getDefaultMiddleware) =>
-      getDefaultMiddleware({ serializableCheck: { warnAfter: 160 } })
-        .prepend(userListener.middleware)
-        .concat(supabaseApi.middleware, spotifyApi.middleware),
+      getDefaultMiddleware({ serializableCheck: { warnAfter: 160 } }),
   });
 };
